Style box model on index.html in styles.css
I positioned heading and unordered list to the center of the page with the help of box model in browser that appears after inspecting.

As for the links in list item elements, I made sure to scale the anchor tag inside of them such that the hit area for links is the entire box that they are contained in.

Also, I gave every box a border of their own, while styling them with colors and adjusting their scaling to maintain a proper proportion between them.